added validateBulkUploadSchema to lib
